+++ packages/client/src/index.js
@@ -1,5 +1,5 @@
 import _ from 'lodash/fp'
-import * as F from 'futil-js'
+import F from 'futil'
 import { flatten, bubbleUp, Tree, encode, decode, isParent } from './util/tree'
 import { validate } from './validation'
 import { getAffectedNodes, reactors } from './reactors'
@@ -65,7 +65,7 @@ export let ContextTree = _.curry(
     let customReactors = reactors
 
     // initNode now generates node keys, so it must be run before flattening the tree
-    dedupeWalk(initNode(extend, types), tree)
+    dedupeWalk(initNode({ extend, types, snapshot }), tree)
     let flat = flatten(tree)
     let getNode = path => flat[encode(path)]
 
@@ -189,7 +189,12 @@ export let ContextTree = _.curry(
             !target.forceReplaceResponse &&
             F.maybeCall(typeProp('shouldMergeResponse', target), target)
           )
-            typeProp('mergeResponse', target)(target, responseNode, extend)
+            typeProp('mergeResponse', target)(
+              target,
+              responseNode,
+              extend,
+              snapshot
+            )
           else {
             target.forceReplaceResponse = false
             extend(target, responseNode)