+++ packages/client/src/node.js
@@ -1,5 +1,5 @@
 import _ from 'lodash/fp'
-import F from 'futil-js'
+import F from 'futil'
 import { Tree, encode } from './util/tree'
 import { runTypeFunction, runTypeFunctionOrDefault, getTypeProp } from './types'
 
@@ -28,19 +28,22 @@ export let internalStateKeys = {
 
 export let autoKey = x => F.compactJoin('-', [x.field, x.type]) || 'node'
 
-export let initNode = _.curry((extend, types, dedupe, parentPath, node) => {
-  runTypeFunction(types, 'init', node, extend)
-  let key = dedupe(
-    node.key ||
-      runTypeFunctionOrDefault(autoKey, types, 'autoKey', node, extend)
-  )
-  extend(node, {
-    ..._.omit(_.keys(node), defaults),
-    ..._.omit(_.keys(node), _.cloneDeep(getTypeProp(types, 'defaults', node))),
-    key,
-    path: [...parentPath, key],
-  })
-})
+export let initNode = _.curry(
+  ({ extend, types, snapshot }, dedupe, parentPath, node) => {
+    runTypeFunction(types, 'init', node, extend)
+    let key = dedupe(
+      node.key ||
+        runTypeFunctionOrDefault(autoKey, types, 'autoKey', node, extend)
+    )
+    extend(node, {
+      ..._.omit(_.keys(node), defaults),
+      // For some reason, type defaults can end up observable in real world apps, so we `snapshot` instead of `_.deepClone`
+      ..._.omit(_.keys(node), snapshot(getTypeProp(types, 'defaults', node))),
+      key,
+      path: [...parentPath, key],
+    })
+  }
+)
 
 // fn: (dedupe: string -> string, parentPath: array, node: object) -> void
 export let dedupeWalk = (fn, tree, { target = {}, dedupe } = {}) => {