+++ packages/client/src/util/futil.js
@@ -1,4 +1,5 @@
-let _ = require('lodash/fp')
+import _ from 'lodash/fp'
+import F from 'futil'
 
 // Sugar for checking non empty intersection
 export let intersects = (a, b) => !_.isEmpty(_.intersection(a, b))
@@ -17,3 +18,23 @@ export let eventEmitter = (listeners = {}) => ({
     }
   },
 })
+
+export let transformTreePostOrder = (next = F.traverse) =>
+  _.curry((f, x) => {
+    let result = _.cloneDeep(x)
+    F.walk(next)(_.noop, f)(result)
+    return result
+  })
+
+export let maybeUpdateOn = _.curry((fn, key, data) =>
+  _.get(key, data) ? F.updateOn(key, fn, data) : data
+)
+
+// Recursively transforming multiple props with supplying produced transformation
+export let deepMultiTransformOn = (props, transformWith) => {
+  let self = _.flow(
+    _.map(maybeUpdateOn(transformWith(x => self(x)))),
+    _.flow
+  )(props)
+  return self
+}